+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.16.2 (2025-07-11)
+
+- Fix `debug::print_vm_stack` which was returning the advice stack instead of the system stack [(#1984)](https://github.com/0xMiden/miden-vm/issues/1984).
+
 ## 0.16.1 (2025-07-10)
 
 - Make `Process::state()` public and re-introduce `From<&Process> for ProcessState`.